Rework Kata09
The `itemPrice` function could be extracted from the `Checkout` and cleaned up a bit, but this is starting to look like something. Right now the calculation is constrained to prices for single items, though. So an offer where you get a discount on product B if you also buy product A is not possible yet.